requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cls-client==1.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.10.1
django-filter==21.1
django-scheduler==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
icalendar==4.0.9
idna==3.3
importlib-metadata==4.8.2
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.5
MarkupSafe==2.0.1
oauthlib==3.1.1
p==1.2.5
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
